refactor(runtime): split native_io.rs into cohesive submodules (§7-8)
Split the 3784-line runtime/native_io.rs (over the 500-line limit) into a
facade plus eleven themed submodules under src/runtime/native_io/. All but
one are under 500 lines; io_handle.rs (740) holds the single ~660-line
native_io_handle method (an indivisible match). The file is a hybrid of
free functions (all originally private) and impl Interpreter methods; the
facade re-exports the cross-submodule free helpers (pub(crate)) and the
impl methods move freely (visibility-only). runtime/mod.rs is untouched
(submodules declared inside the facade). All #[cfg(...)] platform gates
preserved verbatim.

- native_io.rs (19)       - facade: use lines + re-exports
- io_handle.rs (740)      - native_io_handle (single indivisible match)
- canonpath.rs (485)      - canonpath_unix/qnx/cygwin/win32 + volume split
- io_pipe.rs (380)        - native_io_pipe
- io_path_mutate.rs (379) - fs-mutate (symlink/chmod/rename, cfg-gated)
- native_io_path.rs (342) - native_io_path
- io_path_stat.rs (334)   - fs-stat (mode/changed, cfg-gated)
- io_path_lexical.rs (276)- try_io_path_lexical + cwd
- resolve.rs (262)        - resolve_io_path + extension parts
- path_spec.rs (255)      - win32/cygwin path-spec helpers
- io_path_read.rs (228)   - content read + comb
- helpers.rs (140)        - error builders + cfg-gated path_is_* family

Pure relocation. 3 pub(super) methods bumped to pub(crate) (external
callers) + cross-submodule private helpers to pub(crate). The fn multiset
is byte-identical. No behavior change.
